fix(Users): Fixed the external service returning null users.

+++ AidManager.API/ManageTasks/Application/Internal/CommandServices/TaskCommandService.cs
@@ -1,4 +1,5 @@
 ﻿using AidManager.API.Authentication.Domain.Model.Entities;
+using AidManager.API.ManageTasks.Application.Internal.OutboundServices;
 using AidManager.API.ManageTasks.Application.Internal.OutboundServices.ACL;
 using AidManager.API.ManageTasks.Domain.Model.Aggregates;
 using AidManager.API.ManageTasks.Domain.Model.Commands;
@@ -8,7 +9,7 @@ using AidManager.API.Shared.Domain.Repositories;
 
 namespace AidManager.API.ManageTasks.Application.Internal.CommandServices;
 
-public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unitOfWork, IProjectRepository projectRepository, ExternalUserService externalUserService, IProjectCommandService projectCommandService) : ITaskCommandService
+public class TaskCommandService(ITaskEventHandlerService eventHandlerService,ITaskRepository taskRepository, IUnitOfWork unitOfWork, IProjectRepository projectRepository, ExternalUserService externalUserService) : ITaskCommandService
 {
     public async Task<(TaskItem, User)> Handle(CreateTaskCommand command)
     {
@@ -22,6 +23,7 @@ public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unit
             }
             
             var user = await externalUserService.GetUserById(command.AssigneeId);
+            if (user is null) throw new Exception("User not found");
             
             if (user.Role == 0)
             {
@@ -30,12 +32,10 @@ public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unit
             
             var task = new TaskItem(command);
             
-            await projectCommandService.Handle(new AddTeamMemberCommand(command.AssigneeId, command.ProjectId));
-            
-            
-                await taskRepository.AddAsync(task);
-                await unitOfWork.CompleteAsync();
-                return (task, user);
+            await eventHandlerService.HandleAddTeamMember(new AddTeamMemberCommand(command.AssigneeId, command.ProjectId));
+            await taskRepository.AddAsync(task);
+            await unitOfWork.CompleteAsync();
+            return (task, user);
         }
         catch (Exception e)
         {
@@ -55,6 +55,8 @@ public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unit
             throw new Exception($"Project with id {command.ProjectId} does not exist.");
         }
         var user = await externalUserService.GetUserById(command.AssigneeId);
+        if (user is null) throw new Exception("User not found");
+        
         var task = await taskRepository.GetTaskById(command.Id);
 
         if (task is null) throw new Exception("Task not found");
@@ -79,6 +81,7 @@ public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unit
         var task = await taskRepository.GetTaskById(command.Id);
         if (task is null) throw new Exception("Task not found");
         var user = await externalUserService.GetUserById(task.AssigneeId);
+        if (user is null) throw new Exception("User not found");
         await taskRepository.Remove(task);
         await unitOfWork.CompleteAsync();
         return (task, user);
